Auth Upgrades, Error Handling / Validation, Seperation of environments and variables

+++ src/main/java/com/neuroforged/leadsystem/dto/LeadResponseDTO.java
@@ -17,8 +17,8 @@ public class LeadResponseDTO {
     private String trafficSource;
     private Integer monthlyLeads;
     private Double conversionRate;
-    private BigDecimal costPerLead;
-    private BigDecimal clientValue;
+    private Double costPerLead;
+    private Double clientValue;
     private Integer leadScore;
     private String leadChallenge;
     private String clientId;
